+++ lib/runtime/cloudflare.dart
@@ -1,3 +1,4 @@
+export '../src/runtime/cloudflare/export.dart' show defineFetchExport;
 export '../src/runtime/cloudflare/extension.dart'
     show CloudflareRuntimeExtension;
 export '../src/runtime/cloudflare/host.dart';