+++ lib/runtime/vercel.dart
@@ -1,3 +1,4 @@
+export '../src/runtime/vercel/export.dart' show defineFetchExport;
 export '../src/runtime/vercel/extension.dart' show VercelRuntimeExtension;
 export '../src/runtime/vercel/functions.dart'
     show VercelFunctions, VercelRuntimeCache;